+++ app/build.gradle
@@ -53,6 +53,7 @@ kapt {
 dependencies {
     implementation project(":data:api")
     implementation project(":data:model")
+    implementation project(":data:dao")
     implementation project(":util")
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
