+++ app/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/repository/RepoRepository.kt
@@ -1,6 +1,6 @@
 package jp.chau2chaun2.honkot.sample.multimodule.repository
 
-import jp.chau2chaun2.honkot.sample.multimodule.dao.RepoDao
+import jp.chau2chaun2.honkot.sample.multimodule.data.dao.RepoDao
 import jp.chau2chaun2.honkot.sample.multimodule.data.model.Repo
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
