+++ data/dao/src/main/java/jp/chau2chaun2/honkot/sample/multimodule/data/dao/RepoDao.kt
@@ -1,7 +1,7 @@
-package jp.chau2chaun2.honkot.sample.multimodule.dao
+package jp.chau2chaun2.honkot.sample.multimodule.data.dao
 
 import jp.chau2chaun2.honkot.sample.multimodule.data.model.ColorType
-import jp.chau2chaun2.honkot.sample.multimodule.model.OrmaDatabase
+import jp.chau2chaun2.honkot.sample.multimodule.data.model.OrmaDatabase
 import jp.chau2chaun2.honkot.sample.multimodule.data.model.Repo
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -26,7 +26,9 @@ class RepoDao @Inject constructor(
     fun initialize() {
         val random = Random(System.currentTimeMillis())
         val colors = ColorType.values()
-        IntRange(0, INITIALISE_MODEL_COUNT).forEach { index ->
+        IntRange(0,
+            INITIALISE_MODEL_COUNT
+        ).forEach { index ->
             orma.insertIntoRepo(
                 Repo().also {
                 it.index = index
