+++ settings.gradle
@@ -2,4 +2,5 @@ rootProject.name='SampleListDataByMultiModule'
 include ':app'
 include ':data:api'
 include ':data:model'
+include ':data:dao'
 include ':util'
